requirements.txt
Flask==2.3.3
pandas==2.0.3
joblib==1.3.2
numpy==1.24.3
Pillow==10.0.0
scikit-learn==1.3.0
pycryptodome==3.18.0
pydub==0.25.1
